backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
numpy==1.26.2
imutils==0.5.4
scikit-learn==1.3.2
openpyxl==3.1.2
pandas==2.1.3
Pillow==10.1.0
pydantic==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
PyMuPDF==1.23.8
